修正QuickSendTool工具发送过来的文件带中文名编码错误问题

+++ plugins/copyFile.js
@@ -31,6 +31,9 @@ async function copyFilesB(sourcePath, targetPath, blackList) {
 async function deleteFiles(targetPath, cacheFilePath) {
 	for (const cacheFile of cacheFilePath) {
 		const filePath = path.resolve(targetPath, cacheFile);
+        if(!fs.existsSync(filePath)) {
+            continue;
+        }
 		try {
 			await fs.rm(filePath, { recursive: true, force: true });
 		} catch (err) {
